backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
langchain
langchain-groq
groq==0.4.1
pydantic==2.5.0
reportlab==4.0.7
aiofiles==23.2.1
httpx==0.25.0
python-multipart==0.0.6
pymongo==4.15.3
bcrypt==5.0.0
python-jose==3.5.0
email-validator==2.3.0
